# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart ERP & BI System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details with price history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /products/{id}/movements:
    get:
      operationId: getProductMovements
      tags: [products]
      summary: Get stock movements for a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stock movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
  /products/{id}/adjust:
    post:
      operationId: adjustStock
      tags: [products]
      summary: Manual stock adjustment (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustRequest"
      responses:
        "200":
          description: Stock adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        code:
          type: string
        name:
          type: string
        category:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        currentQuantity:
          type: number
        safetyStock:
          type: integer
        costPrice:
          type: number
        salePrice:
          type: number
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        stockStatus:
          type: string
          enum: [available, low, out]
        createdAt:
          type: string
          format: date-time
      required: [id, companyId, code, name, currentQuantity, safetyStock, costPrice, salePrice, stockStatus, createdAt]
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            priceHistory:
              type: array
              items:
                $ref: "#/components/schemas/PriceHistoryEntry"
            movements:
              type: array
              items:
                $ref: "#/components/schemas/StockMovement"
          required: [priceHistory, movements]
    PriceHistoryEntry:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        supplierId:
          type: integer
          nullable: true
        supplierName:
          type: string
          nullable: true
        oldPrice:
          type: number
        newPrice:
          type: number
        changePercent:
          type: number
        recordedAt:
          type: string
          format: date-time
      required: [id, productId, oldPrice, newPrice, changePercent, recordedAt]
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: number
        reference:
          type: string
          nullable: true
        userId:
          type: integer
          nullable: true
        userName:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, productId, type, quantity, createdAt]
    StockAdjustRequest:
      type: object
      properties:
        quantity:
          type: number
        type:
          type: string
          enum: [in, out]
        notes:
          type: string
          nullable: true
      required: [quantity, type]
    CreateProductRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        category:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        safetyStock:
          type: integer
        costPrice:
          type: number
        salePrice:
          type: number
        supplierId:
          type: integer
          nullable: true
      required: [code, name, safetyStock, costPrice, salePrice]
...