# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart ERP & BI System API
 * OpenAPI spec version: 0.1.0
 */
import type { StockMovementType } from "./stockMovementType";

export interface StockMovement {
  id: number;
  productId: number;
  type: StockMovementType;
  quantity: number;
  reference?: string | null;
  userId?: number | null;
  userName?: string | null;
  notes?: string | null;
  createdAt: Date;
}